refactor: move API code for FFs to Core
Ref. metriport/metriport-internal#2840
