+++ packages/api/src/external/carequality/patient.ts
@@ -1,3 +1,4 @@
+import { isDemoAugEnabledForCx } from "@metriport/core/command/feature-flags/domain-ffs";
 import { Patient, PatientExternalData } from "@metriport/core/domain/patient";
 import { toIheGatewayPatientResource } from "@metriport/core/external/carequality/ihe-gateway-v2/patient";
 import { MedicalDataSource } from "@metriport/core/external/index";
@@ -8,18 +9,17 @@ import { OutboundPatientDiscoveryReq } from "@metriport/ihe-gateway-sdk";
 import { errorToString } from "@metriport/shared";
 import dayjs from "dayjs";
 import duration from "dayjs/plugin/duration";
+import { createAugmentedPatient } from "../../domain/medical/patient-demographics";
+import { resetScheduledPatientDiscovery } from "../hie/reset-scheduled-patient-discovery-request";
 import { makeIHEGatewayV2 } from "../ihe-gateway-v2/ihe-gateway-v2-factory";
 import { makeOutboundResultPoller } from "../ihe-gateway/outbound-result-poller-factory";
 import { deleteCQPatientData } from "./command/cq-patient-data/delete-cq-data";
+import { updatePatientDiscoveryStatus } from "./command/update-patient-discovery-status";
 import { createOutboundPatientDiscoveryReq } from "./create-outbound-patient-discovery-req";
 import { gatherXCPDGateways } from "./gateway";
 import { PatientDataCarequality } from "./patient-shared";
-import { updatePatientDiscoveryStatus } from "./command/update-patient-discovery-status";
-import { getCqInitiator, isCqEnabled } from "./shared";
 import { queryDocsIfScheduled } from "./process-outbound-patient-discovery-resps";
-import { createAugmentedPatient } from "../../domain/medical/patient-demographics";
-import { resetScheduledPatientDiscovery } from "../hie/reset-scheduled-patient-discovery-request";
-import { isDemoAugEnabledForCx } from "../aws/app-config";
+import { getCqInitiator, isCqEnabled } from "./shared";
 
 dayjs.extend(duration);
